pyproject.toml
[tool.poetry]
name = "alita-mini"
version = "0.1.0"
description = ""
authors = ["hellozhaojian <290169713@qq.com>"]
readme = "README.md"
packages = [{include = "alita_mini"}]




[tool.poetry.dependencies]
#python = "^3.9"
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
requests = "^2.31.0"
openai = "^0.27.8"
pyyaml = "^6.0"
pydantic = "1.10.10"
pdfplumber = "^0.9.0"
tiktoken = "^0.4.0"
langchain = "^0.0.222"
jinja2 = "^3.1.2"
python-dotenv = "^1.0.0"
dash = "^2.11.1"
streamlit = "1.22.0"
bson = "^0.5.10"
fastapi = "^0.99.1"
motor = "^3.2.0"
uvicorn = "^0.22.0"
redis = "^4.6.0"
pymongo = "^4.4.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pytest-cov = "^4.1.0"
coverage = "^7.2.7"
ipykernel = "^6.23.3"


[build-system]
requires = [
  "python >= 3.9,< 3.9.7 || >3.9.7,< 4.0",
  "poetry-core"
]
build-backend = "poetry.core.masonry.api"



[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
  \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | venv
)/
'''

[tool.isort]
profile = "black"
known_third_party = ["django", "numpy", "pandas"]
known_first_party = ["alita_mini"]

